Change TotalObjects type from int32 -> int64
The original change was int -> int32, which on 64bit machines is a
downgrade that I overlooked.  Raising here to int64 to maintain the
size.

+++ cmd/tempo-cli/cmd-list-compactionsummary.go
@@ -61,9 +61,9 @@ func displayCompactionSummary(results []blockStats) {
 		sizeSum := uint64(0)
 		sizeMin := uint64(0)
 		sizeMax := uint64(0)
-		countSum := int32(0)
-		countMin := int32(0)
-		countMax := int32(0)
+		countSum := int64(0)
+		countMin := int64(0)
+		countMax := int64(0)
 		countBloomShards := 0
 
 		var newest time.Time
