+++ tempodb/encoding/vparquet4/block_iterator_test.go
@@ -32,7 +32,7 @@ func TestRawIteratorReadsAllRows(t *testing.T) {
 	require.NoError(t, err)
 	defer iter.Close()
 
-	actualCount := int32(0)
+	actualCount := int64(0)
 	for {
 		_, tr, err := iter.Next(context.Background())
 		if tr == nil {
